Create FINAL PROJECT
Hello! My project's name is 'A slightly less tougher version of a Scientific Calculator'. It is a calculator, but with
comparatively fewer functions. Most of the time, we come across several situations, especially in our school life, when 
we are stuck doing our maths homework, but we only have a simple calculator. At such instants, calculating powers of numbers, 
and other functions becomes very tedious. That is why i have introduced various functions, which will help make our 
calculations much, much simpler.
The functions included are :
1.Addition
2.Subtraction
3.Multiplication
4.Division
5.Square Root
6.Power
7.Square
8.Cube
9.Reciprocal
10.Power to Reciprocal
11.10 to the Power of....
12.Factorial
13.Remainder
14.Logarithms
15.Modulus
16.Sin
17.Cos
18.Tan
19.Cosec
20.Sec
21.Cot